Reorganiza arquivos
Reorganiza os arquivos separando em codigo de testes e de producao
Atualiza teste  de modelo para rodar setUp e tearDown
