+++ Makefile
@@ -1,9 +1,14 @@
 venv_path ?=.venv
-source_venv =. $(venv_path)/bin/activate &&
+source_venv =. $(venv_path)/bin/activate && 
 
 init:
-	test -d $(venv_path) || python3 -m venv $(venv_path)
-	$(source_venv) pip install -r requirements.txt
+	pip install -r requirements.txt
 
-run: init
-	$(source_venv) gunicorn app:app
+activate:
+	test -d $(venv_path) || python3 -m venv $(venv_path)
+
+run: activate
+	$(source_venv) gunicorn app:app --chdir agenciacompromisso
+
+test: activate
+	$(source_venv) python3 -m unittest discover --verbose test
